+++ NEWS.md
@@ -1,4 +1,5 @@
 # zoomGroupStats 0.1.9000
+* Added `flattenSelf()` and `batchFlattenSelf` to create better conversational turns in transcript and chat files
 * Added `getVideoDateTimeMeta()` function to extract the creation timestamp for the recording from the video metadata
 * Added `imageDetectTimeStamp()` function to extract the recording time from a still frame of a Zoom video